modified makefiles and staged day3 fixes

+++ day3/day3_part2.cpp
@@ -47,7 +47,7 @@ int main(int argc, char * argv[])
     width = co2[0].size();
     
     for(int bit = 0; bit < width && co2.size() > 1; bit++) {
-        int c0 = 0;
+        size_t c0 = 0;
         char lc_rev;
         for(string number : co2)
             if(number.at(bit) == '0') c0++;
@@ -61,7 +61,7 @@ int main(int argc, char * argv[])
 
     // copy - pasted from above and tweaked due to laziness
     for(int bit = 0; bit < width && oxygen.size() > 1; bit++) {
-        int c1 = 0;
+        size_t c1 = 0;
         char mc_rev;
         for(string number : oxygen)
             if(number.at(bit) == '1') c1++;
